Refactor email draft handling to thread-level draft management
